# book-network-ui/src/app/services/models/book-responce.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookResponce {
  archived?: boolean;
  authorName?: string;
  cover?: string;
  id?: number;
  isbn?: string;
  owner?: string;
  rate?: number;
  sharable?: boolean;
  synopsis?: string;
  title?: string;
}