# book-network-ui/src/app/services/models/page-responce-book-responce.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BookResponce } from '../models/book-responce';
export interface PageResponceBookResponce {
  content?: Array<BookResponce>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}