# book-network-ui/src/app/services/models/page-responce-borrowed-book-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BorrowedBookResponse } from '../models/borrowed-book-response';
export interface PageResponceBorrowedBookResponse {
  content?: Array<BorrowedBookResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}