# book-network-ui/src/app/services/models/page-responce-feedback-responce.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FeedbackResponce } from '../models/feedback-responce';
export interface PageResponceFeedbackResponce {
  content?: Array<FeedbackResponce>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}